pyproject.toml
[tool.poetry]
name = "langchain-client"
version = "0.1.0"
description = ""
authors = ["Robert Todea <robert.todea@mai.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.10"
langchain-cli = "^0.0.21"
langsmith = "^0.1.19"
langchain-openai = "^0.0.8"
python-dotenv = "^1.0.1"
ipython = "^8.22.2"
beautifulsoup4 = "^4.12.3"
google-generativeai = "^0.4.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
